+++ CHANGELOG.md
@@ -4,6 +4,16 @@ that can be found in the LICENSE file. -->
 
 # Changelog
 
+## 3.7.0
+
+### New features
+
+- Add `preferredFlashMode`, allowing users to choose which flash mode is preferred when first using the camera. (#158)
+
+### Improvements
+
+- Allow flash modes failed to switch and can move on to next when switching. (#156)
+
 ## 3.6.5
 
 ### Fixes